+++ main.cpp
@@ -66,7 +66,7 @@ int main()
         DepthSensorModel KinectModel("../../resources/KinectModel.xml");
         KinectModel.computeCov(80, 360, 0.5837, cov);
         Eigen::Vector3d vec;
-        KinectModel.getPoint(377.177, 112.906, 6.468, vec);
+        KinectModel.getPoint((long int)377.177, (long int)112.906, 6.468, vec);
 
         // create objects and print configuration
         cout << "Current grabber: " << grabber->getName() << std::endl;
